Replaced megasd
The megasd has been rebuilt.

(1) Reduced the circuit size
(2) Added low-speed access mode.

The low-speed access mode is for establishing a connection with an SD/MMC card. In order to support this, the software side must also be supported.
If nothing is done, the access will be the same as before.